src/main/java/ladder/domain/Ladder.java
package ladder.domain;

import java.util.List;
import java.util.Objects;
import java.util.stream.Collectors;
import java.util.stream.IntStream;

public class Ladder {
    private final List<Line> ladder;

    public Ladder(int height, int length, LineGenerator lineGenerator) {
        this.ladder = IntStream.range(0, height)
                .mapToObj(i -> lineGenerator.create(length))
                .collect(Collectors.toList());
    }

    public List<Line> ladder() {
        return ladder;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Ladder ladder1 = (Ladder) o;
        return Objects.equals(ladder, ladder1.ladder);
    }

    @Override
    public int hashCode() {
        return Objects.hash(ladder);
    }
}
